Fixed a bug where Content-Type was incorrectly reported
Because it was reported using `String.prototype.length`, which is the
number of characters, not the number of bytes.  Apparently JavaScript
doesn't have a build-in way to calculate the true number of bytes in a
string.

My bad.  I shouldn't have assumed that JavaScript strings had 1-byte
characters.  This isn't C dude.